Refactor for readability, consistency, reliablity
For readability, refactor the main function to a short method. Using
a structure `info_t` to pass the variables by. Added "echo" to inform
users about the validation results.

For consistency, remove the redundant malloc in OPT:append().
Because it will cause the searched/not searched strings have not/have
the character '\n'. Instead, removing the character '\n' in file_align.

For reliablity, it should free every element of linked-list. In ORIG,
the head of the list is empty. After appending, the head should be
removed. In OPT, the entry_pool's size is equal to the input. Because of
the $(NUM_THREAD) empty heads, it causes the lackage of output.